replace -D_POSIX_C_SOURCE=1 by -D_POSIX_C_SOURCE=200112L
POSIX_C_SOURCE=1 only specifies that it is posix compliant source we
want to compile, but it does not define which posix standard we adhere
to.

On (open)solaris, this fails, as the compiler assumes POSIX.1-1999
(probably comming from -std=c99). Setting POSIX_C_SOURCE to 200112L
switches the internas to the POSIX.1-2001 standard.

This solved issue #45