Allow to forcibly build pacrunners as module.
The default is to build the pacrunner internal to libproxy if only one
pacrunner is to be built.

Using -DBIPR{ON|OFF} it is now possible to override this and have a
pacrunner built as module even if there is only one to be built.

In case of multiple pacrunners, BIPR is being ignored and forcibly set
to OFF.

Fixes issue 172.