Fix test #10
Fix the way we detect if get_pac correctly aborted on our server trying
to send a > 102400 byte long stream (which we do to avoid malicious pac
files/wpad server to buffer overflow us).

Replace assert calls with proper error handling, which avoids the server
from hanging (the write command to send the quit was wrapped in
assert()).

When not building for a Debug release with CMAKE, the -DNDEBUG is being
specified, which in turn disables the assert macro.

Fixes bug 189.