bindings: perl: Do not link against libperl.
As the binding is always running in context of perl, the
perl libraries are guaranteed to be in memory and we do not
have to take care of those.

This allows the perl bindings to be built against a different
version of perl as it will be finally run with (perl keeps on
installing the library in a versioned directory, hence on each
perl upgrade, a rebuild is needed in order to link against the
new location of the library).