pxgsettings: g_signal_connect does not emmit changed signals
In recent versions of GLIB, it is mandatory to FIRST connect to the changed:: signal and then read values.
Without reading the values, it seems there is an initialization missing to start the change monitoring.

So: we now first connect the signal handler, directly followed by reading all the settings we care about.