Revert "bindings: perl: Do not link against libperl."
This reverts commit 5077b7563369a89d906fa32c040bb953c3b628bb.

The original change breaks linking with -Wl,--no-undefined, which is the
default on distros such as Fedora and Debian, and it also makes some
sanity tests on FreeBSD to fail since the code uses Perl symbols but
does not link against libperl.so. There does not seem to be a big
benefit in building the binding with a Perl version other than the
installed on, especially since most distros simply rebuild everything
when switching Perl versions.