bindings: perl: Add an option to explicitly link against libperl.so.
The Perl binding uses symbols from libperl.so, but since it is always
run after libperl.so itself is loaded, not linking against it does not
cause the problems that would normally occur.

Doing so was mandatory since 5077b75 ("bindings: perl: Do not link
against libperl"), which was catering for distros which install the
binding into vendor_perl and try not to rebuild all their Perl packages
on every Perl update. The flip-side is that it breaks things for other
groups: distros that install the binding into a versioned Perl
directory and allow different Perl versions to be installed
concurrently (for example, FreeBSD), or distros that pass
"--no-undefined" to the linker.

Provide an option, PERL_LINK_LIBPERL (whose default depends on the value
of PERL_VENDORINSTALL), to make the behavior configurable.