python: Avoid a crash on 64-bit systems
Annotate the return type of px_proxy_factory_new() to be void *, as otherwise
int is assumed. This works fine on 32-bit systems, where void * and int are the
same width, but is invalid on 64-bit (Linux) systems.

Additionally, annotate the argument type of free() and px_proxy_factory_free()
to be void * to match.

https://code.google.com/archive/p/libproxy/issues/146