Use different name when checking for 2nd set NetworkManager packages
Seems different to Autoconf, cmake does keep some cache around and does
not transparently reuse the name fully to make a complete test. This for
one resulted in old versions of NetworkManager being queried, yet the final
build was disabled. If cmake was ran a 2nd time the cache was refreshed and
the NetworkManager module being built.

By using a 2nd name we aboid the internal confusion for cmake.

Fixes issue #53