Keep visibility of non-px_* symbols
Symbols other than px_* are still used by modules, so they have to be visible
as well. Unfortunately the version script can't deal with C++ classes properly,
it's not possible to match the vtable, typeinfo, etc. without resorting to the
platform specific mangling.

As a workaround, keep the default visibility of symbols which aren't explicitly
mentioned, this means that while users of private API wouldn't be caught
anymore, everything continues to work as usual, just with px_* being versioned.