python: Convert URLs to/from UTF-8 under Python 3
In Python 3, strings are Unicode (UTF-16 or UTF-32 depending on build
options), and are passed to ctypes function wrappers as such.
Convert the input URL to a bytestring for processing by libproxy, and
convert the results back into (Unicode) strings by decoding them.

From the pacrunner_webkit module, it appears that libproxy expects
strings to be in UTF-8 (like GLib/GTK), so use that encoding for
bytestrings.

Fixes: #65