gsettings: Always read at least the initial values of all keys
Previously, the gsettings (gnome3) module would wait until it had read
at least one key/value pair for each GSettings schema requested on the
command line. However, GSettings schemas often contain more than one
key. With a relatively elaborate proxy configuration, we don't
necessarily read all of them in one read() from the pipe, resulting in
the full configuration not being read until later - but for proxy(1),
"later" is too late, because it exits after the first query.

To resolve this, add a way for pxgsettings to signal that it has
written all the initial values (a blank line), and wait for that.

Fixes: #116
Signed-off-by: Simon McVittie <smcv@debian.org>