Disable mozjs extension by default
mozjs doesn't have a stable API, so distros wind up carrying a bunch of
patches to make it work with newer mozjs. Crashes ensue if it gets
linked into gjs applications and the versions don't match. There's
really no benefit to using this over the WebKit extension, so let's
nudge distros towards using that instead.

Distros: don't wait for this commit, go ahead and build with
-DWITH_MOZJS=OFF today to use the WebKit backend instead.