Specify library version more completely
If the library version is not specified, meson installs the shared
object using only the soversion, and this causes ldconfig to emit a
warning that the shared object is not a symlink. Specifying a version
that matches the project version means that version is used for the
shared object, and the soversion-named file is then a symilnk to that,
which is common across other libraries, and avoids the warning from
ldconfig.

https://bugzilla.redhat.com/show_bug.cgi?id=2247508