Add curl option to the generated config for backend instead
It is only used there, and the project root is added to the list of
include directories last. This means that a config.h may inadvertedly be
included from a completely different place when building libproxy as a
subproject. The backend directory, where the generated
px-backend-config.h ends up is the first include path in the list.

This commit removes config.h entirely, as the only other thing in it,
the generated PACKAGE_VERSION define, is not used anywhere