tests: Copy pxProxyFactory correctly
The second argument to g_boxed_copy() is meant to be a pointer to the
boxed instance itself, not a pointer to a pointer to the boxed instance.
For example, `g_boxed_copy (G_TYPE_STRV, environ)` is correct (equivalent
to g_strdupv() in this case), but `g_boxed_copy (G_TYPE_STRV, &environ)`
would be incorrect. The same principle applies here.

The test passed despite this because self->pf and
&self->pf happen to both be pointer-sized, and as a result of
https://github.com/libproxy/libproxy/issues/280, the data that is pointed
to was copied byte-by-byte but not otherwise used; but this will no
longer be true when #280 is fixed, so the test needs fixing first.

Signed-off-by: Simon McVittie <smcv@debian.org>